requirements.txt
fastapi[standard]
vosk
piper-tts
